+++ src/display/theme/colours.rs
@@ -276,14 +276,9 @@ impl RgbColours {
         Self::theme().entry_directory.colour
     }
 
-    /// Returns the theme colour for summary numbers.
-    pub(crate) fn summary_number() -> Colour {
-        Self::theme().summary_number.colour
-    }
-
     /// Returns the theme colour for summary text.
-    pub(crate) fn summary_text() -> Colour {
-        Self::theme().summary_text.colour
+    pub(crate) fn summary() -> Colour {
+        Self::theme().summary.colour
     }
     pub(crate) const ZESTY: Color                    = Color::Rgb(248, 248, 148);
     pub(crate) const LILLIPUTIAN_LIME: Color         = Color::Rgb(137, 227, 81);